fix(archives): contain extraction paths and plan collision-free output names
Archive extraction now resolves every entry through a shared containment helper
that keeps outputs lexically beneath the caller-owned root, applied across WAD,
PRE, compressed PRE, PKR, PAK, BON, DDX, CUT, QZip, and disc images. Batch output
naming is planned up front for mesh, bitmap, and script conversion so
platform-suffixed inputs can no longer overwrite one another.

Work produced in the parallel Codex session; committed here after verifying
both targets build and the suite runs (see the batch report for known failures).

+++ src/NeversoftMultitool/Core/Formats/DiscImage/GcmFileSystem.cs
@@ -33,7 +33,7 @@ public static class GcmFileSystem
         var fstOffset = BinaryPrimitives.ReadUInt32BigEndian(header);
         var fstSize = BinaryPrimitives.ReadUInt32BigEndian(header[4..]);
 
-        if (fstOffset == 0 || fstSize == 0 || fstOffset + fstSize > stream.Length)
+        if (fstOffset == 0 || fstSize == 0 || (long)fstOffset + fstSize > stream.Length)
             throw new InvalidDataException("GCM FST offset/size invalid.");
 
         var fst = new byte[fstSize];
